+++ multi-token-edit/multi-token-edit.mjs
@@ -50,6 +50,7 @@ globalThis.MassEdit = {
   importSceneCompendium,
   openPresetBrowser,
   FileIndexer: FileIndexerAPI,
+  sceneNotFoundMessages: [],
 };
 
 // Initialize module